refactor: extract shared /new layout for fonts and CSS vars
Add app/new/layout.tsx + layout.module.css to own the 4 Google fonts,
CSS custom properties, and flex-column root wrapper shared by all /new
routes. Strip the duplicated font imports, fontVars, and .root block
from app/new/page.tsx and app/new/store-list/page.tsx.
